Light stemcells use different stemcell ID's
The AWS CPI supports an optimization called "light stemcells". A light
AWS stemcell is not a full disk image, but rather just a reference to an
AMI. The AWS CPI has been conflating an AMI ID with a stemcell ID on
AWS: when the BOSH director creates a stemcell, the identifier returned
by the AWS CPI was the AMI ID. In the light stemcell case, the stemcell
ID returned was also the referenced AMI ID. This means that when BOSH
director tries to delete the stemcell later, the CPI won't have a good
way to tell the two apart, and hence could only always try to deregister
the AMI, which is undesirable.

This commit changes how light stemcells are identified and behave.

- Stemcell ID for light stemcells now end id " light"
- AWS CPI will no longer attempt to deregister the AMI when deleting a
  light stemcell.

[fixes #61408126]