Test for the value of cloud config key not existance
For better or worse the Bosh paradigm is to consider configuration keys
to be missing if they have a nil value not whether the key is present in
the configureation. Given that, the cpi template will put in
access_key_id and secret_access_key regardless of whether aws is
configured with static credentials source.  This means we cannot test
for the existance of the key but rather the nilness of it.

[#103338736](https://www.pivotaltracker.com/story/show/103338736)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>