Create an end-to-end test for configuration with an IAM instance profile.
Introduce a new job in the pipeline, iam-instance-profile, that
contains tasks for deploying a new director with IAM instance profile
configurations and then deploying a release after. We are using the same
network configurations as ubuntu which means the jobs now need to run
serially. This will be fixed when we setup a new test environment.

In the cloudformation job, we now create an IAM instance profile.
Additionally, did some general refactoring of the template.

[#103431370](https://www.pivotaltracker.com/story/show/103431370)

Signed-off-by: Aaron Hurley <ahurley@pivotal.io>