Breaks Bosh integration section out of cloudformation template
We now have a template that we call 'primary' this manages a stack for
the BATs & Lifecycle suites.  The other template we call
'bosh-integration' and it is used to manage the Bosh Integration suite.
Each template is run from it's own concourse job.  The bosh integration
suite will use this new stack and can be configured to use a different
AWS account than the primary stack

Also:
 - removed the keypairname input from the templates as they are no
   longer used anywhere

[#104220528](https://www.pivotaltracker.com/story/show/104220528)

Signed-off-by: Aaron Hurley <ahurley@pivotal.io>