Adds support for china regions when using AWS::EC2 SDK
the china region is required to be in a '.cn' TLD so we need some switch
logic when composing endpoints.

We also hid the helper to compose endpoint as it should not be exposed
on the Cloud class

Finally we stopped defaulting the passed in options['aws']['region'] key
as we have validation logic that requires that key to be passed in.

[#104220528](https://www.pivotaltracker.com/story/show/104220528)

Signed-off-by: Beyhan Veli <beyhan.veli@sap.com>