Implement simple download test for china region.
A setup and a teardown job were added in order to allow the download
test itself to run individually in order to support future Concourse
metrics. Also, the check for running instances was refactored into a
new file in order to allow the cloudformation task to be used for
any CloudFormation template that does not take in any parameters.

[#104220528](https://www.pivotaltracker.com/story/show/104220528)

Signed-off-by: Edwin Xie <exie@pivotal.io>