Fixes private key name for downloader-cn job
The contents needs to correspond to the public key used by the instance
we ssh into.  We now use a separate public key called `downloader-cn`
for this instance so we wanted to name the private key in the same way
though the actual name of the private key file doesn't matter this
change is only to help clarify what public key we expect to be on the
running instance.

[#106224138](https://www.pivotaltracker.com/story/show/106224138)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>