Verify raw ephemeral disks with an end to end test.
Continued to refactor the end-to-end testing suite so that it follows
RSpec conventions. A new release has been added to ci/assets which will
be used for the e2e testing. We decided to create a BOSH errand for each
test and the run.erb within each errand will define the actual test
that's being ran. This will help provider a clearer context for what is
being tested within the e2e suite. Since we did not want to upload a
release tarball, the run-e2e.sh creates and uploads the release prior to
the specs running. This allowed us to remove the dummy release from the
pipeline. Lastly, created a separate key for the e2e tests to use.

[#104248646](https://www.pivotaltracker.com/story/show/104248646)

Signed-off-by: Jan von Loewenstein <jan.von.loewenstein@sap.com>