Use HVM stemcells for gating the pipeline
Previously, only the Ubuntu BATs used the HVM stemcell, and the Centos
BATs used the non-HVM stemcell, which may account for the build
stability differences between the Ubuntu BATs and the Centos BATs. This
commit sets it so that both BATs that gate the pipeline use the HVM
stemcell while additionally adding two jobs that test using the non-HVM
stemcell in order to investigate the stability of builds with non-HVM
stemcells.

[#107047114](https://www.pivotaltracker.com/story/show/107047114)

Signed-off-by: Edwin Xie <exie@pivotal.io>