Providing more awareness to what the e2e tests are covering.
The e2e setup-director includes configurations for an s3 blobstore. We
don't have a great way of verifying this at the job/errand level right
now so we are just updating the spec name (if 'bosh deploy' succeeds
then it is properly configured). There is a story in our
backlog which should help us improve this coverage.

[#104248646](https://www.pivotaltracker.com/story/show/104248646)

Signed-off-by: Edwin Xie <exie@pivotal.io>