Ensure the light stemcell building vm is destroyed after build
Regardless of success/failure of build.  Note: we cribbed the on_exit
trap code from a bosh-core helper script with some modifications.  We
simplified the add_on_exit function to push on to the array rather than
unshift onto it.  This causes all on_exit items to be executed in the
order they were added to the array instead of the reverse order.  We
also simplified the implementation of this function from it's original

[#104220500](https://www.pivotaltracker.com/story/show/104220500)

Signed-off-by: Thinh Nguyen <thinh.nguyen@emc.com>