Omit "IP address" warning on rate limit exception
Instance creation was incorrectly logging a warning of "IP address was in
use" when the API request had failed due to rate limiting. This is because
2953c67 extended the existing `Bosh::Common.retryable` block without
changing the condition that warnings were logged on.

Fix this by only logging a warning if the exception indicates that the IP
address is invalid. We no longer need to check the number of tries, because
`error` is always `nil` on the first loop.

The tests feel slightly brittle and laboured in that they are matching the
existence and non-existence of some text. But it does seem like it's worth
testing for.