check-for-running-instances task considers stack-created instances
Previously, certain stacks that create an instance (such as end2end)
would cause the check-for-running-instances task to fail, making the
recreation of a stack a manual process. Now we check an instance's tag
to see if it includes a tag with the value of its stack-name. When
Cloudformation creates an instance, it creates a
'aws:cloudformation:stack-name' tag with the value of the stack name.

[#106805970](https://www.pivotaltracker.com/story/show/106805970)