Empties a bucket's contents before attempting to delete the stack
In the teardown-infrastructure task, the Cloudformation command of deleting a stack will fail if a bucket exists and is not empty. Now we will check to see if a bucket exists and if it has contents, if so, it will be emptied before attempting to delete the Cloudformation stack. The End2End tests require a 'stack_prefix' param while the others do not (since a bucket is not needed for those).

[#109367456](https://www.pivotaltracker.com/story/show/109367456)