Fix serial_groups so bats and radiator jobs can run concurrently.
Putting the recreate-infrastructure, bats, and radiator jobs in the same
serial_group made it so that bats and radiator jobs could not run at the
same time which is undesired. This fix should now allow the radiator and
bats jobs to run at the same time but neither can run when
recreate-infrastructure is running and vice-versa.

[#102535032](https://www.pivotaltracker.com/story/show/102535032)