Use region in certify stemcell pipeline config.
Fix an issue with the certify stemcell pipeline where we were not
passing in a host, so it was attempting to authenticate with the default
US region using China credentials. Now we're specifically passing in a
region instead of a host, and we're fixing the flag use_iam to false so
it uses the static credentials already in place.

The reason behind this change is that AWS China only supports V4
signing, and that V4 signing only works with region configuration and
not host/endpoint.

[#110003756](https://www.pivotaltracker.com/story/show/110003756)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>