Refactor use_iam to take an explicit value.
Previously, it was optional and implicitly set to false when not passed
in and therefore initially valued as nil. Now, it's required and passed
in as false explicitly.

NOTE: At the current state, all the use cases for this flag are false,
so we may want to follow up with another refactoring and clean this up,
unless we're planning on adding a new ci job to exercise the iam profile
config soon enough.

[#110003756](https://www.pivotaltracker.com/story/show/110003756)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>