Makes blobstore configuration more explicit for director setup
Introduces new required param `use_s3_blobstore` for the setup-director
task.  If true it will setup the director's blobstore to use the same
aws region as the director, otherwise it will setup a dav blobstore.

We are also turning on env_or_profile for the director's cpi in the
stemcell certification pipeline as that is the only place we are
testing the cpi's ability to use the IAM instance profile.

[#110003756](https://www.pivotaltracker.com/story/show/110003756)

Signed-off-by: Wagner Camarao <wcamarao@vmware.com>