Extracts common logic for ebs volume creation
The constrants when specifying the ebs ephemeral volume mounted on a vm
when it is created in `#create_vm` are similar to those needed for the
ebs persistent volumes created in `#create_disk`.  Each implementation
was verifying those constraints differently for no good reason.

In an attempt to consolidate and normalize the rules around ephemeral vs
persistent volumes we have created a VolumeProperties class to do
validations and InstancesCreatePresnter, VolumesCreatePresenter classes
to prepare the options for the EC2 sdk methods that consume those
properties.

[#110211692](https://www.pivotaltracker.com/story/show/110211692)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>