Job supports the property aws.max_retries
Add the property `aws.max_retries` to the spec and the `cpi.json`
template. This option is passed to the AWS Ruby SDK library:

http://docs.aws.amazon.com/AWSRubySDK/latest/AWS/Core/Configuration.html#max_retries-instance_method

it allows specify the maximum number of times AWS service errors (500) and
throttling errors should be retried. There is an exponential backoff in
between retries, so the more retries the longer it can take to fail.

Increasing this number would mitigate the errors caused Request Rate Limit
from AWS.