Do not retry in ResourceWait on rate limit
As per #23, the logic to retry when there is a RequestLimitExceded is
in the client library code, so it is not needed at an upper level.

In this case we only remove the default error added as errors to retry.

Note that currently this leaves some cases, like `for_volumes`,
`for_snapshot` and `for_sgroup` without an actual valid criteria to
retry as no errors are passed to `Bosh::Retryable`. I prefer to keep
it like this to minimise the number of changes, keep consistency across
handling resources and allow easily add other errors in the future.