Create an existing instance with existing disk for the "existing disk" test
Previously, this test relied on the fact that the CPI had a globally
configured availability zone to create an existing disk before creating
an instance. With the removal of this global AZ, the AZ for the disk
did not match the AZ of the subnet in which a created instance lives.

This test now runs through the recreate VM workflow to more accurately
portray the use case for the CPI handling existing disks.

[#116934273](https://www.pivotaltracker.com/story/show/116934273)
Signed-off-by: Lyle Franklin <lfranklin@pivotal.io>

Signed-off-by: Edwin Xie <exie@pivotal.io>