Ensure light stemcells contain the " light" suffix
A bug was introduced where the CPI would try to delete the AMIs of
light stemcells. Typically this is a non-issue with proper access
control, plus the delete stemcell method swallows authorization errors.

TODO: review logic around "light" handling and remove the swallowing of
the authorization errors, with improved coverage.

[#117964793](https://www.pivotaltracker.com/story/show/117964793)

Signed-off-by: Lyle Franklin <lfranklin@pivotal.io>