Add support for updating AWS routing table entries
The `advertised_instance_routes` cloud property on resource pools/vm
types has been added to allow manifests to connect routes on AWS
Routing Tables with specific instances, if you need to route traffic
through a BOSH deployed VM (like in the case of a site-to-site VPN).

If using this, it is recommended to only have a single instance of a
single job of that resource pool, to avoid confusion. If there are
multiple instances/jobs in the resource pool, the last one processed
will be the node assigned to the route.

Datastructure of `advertised_instance_routes` is as follows:

```
cloud_properties:
  advertised_instance_routes:
  - table_id:    rt-abcdef123
    destination: 10.0.0.0/16
  - table_id:    rt 12345abcd
    destination: 10.0.0.0/16
```

This allows the manifest to specify that multiple routing tables
are updated to send the destination route(s) to it. Once removed
from the manifest, the routes will remain, but once the VM is recreated,
the route will become a black hole in the AWS routing table, and cease
to work.