Use hostname short name as envname for terraform
We pass the output of $(hostname) as the value for the `env_name`
in terraform, which us used in some resources names.

But the command `hostname` might return a FQDN of the hosts with dots,
and some might not like having dots in the name, like the ALB:

    * aws_alb_target_group.default: Error creating LB Target Group:
    ValidationError: Target group name 'keypro.local-local-integration' can
    only contain characters that are alphanumeric characters or hyphens(-)
    status code: 400, request id: b698d1ea-c4c1-11e7-ad82-5fca118e6acc

We would get the shortname of the host with `hostname -s` instead.