Add AWS session_token param to the bosh release
We want to be able to optionally pass the session_token to the cpi,
deployed in the bosh release, so we add it to the spec and to the
cpi.json template. By default the attribute is nil and it is populated
as nil in the cpi.json file.

We update the unit test of the cpi.json template accordingly, with
an additional test case for a populated value.