Optionally use session_token for AWS STS creds
We want to be able to use AWS STS creds, which require an
additional session_token to be passed.

The constructor Aws::Credentials already accepts this additional
3rd parameter. By default is nil, which would mean use normal
credentials.

In this commit read the attribute session_token from the config, which
might be nil if it is not defined, and we pass it to the costructor
of Aws::Credentials.