Improve NVMe support for newer instance types
Seems like this information is not otherwise available from the AWS API,
so we statically list supported families/types here. At some point NVMe
may be the default and we might be able to invert the logic. We will
still need to explicitly add instance types as they are introduced.

[#163595513](https://www.pivotaltracker.com/story/show/163595513)