`sudo` the one executable, not the entire script
Instead of running the entire `stemcell-copy` as root, which is
unnecessary, we constrain root privileges to the single `dd` command.

This lays the groundwork for allowing non-privileged users to run `bosh
create-env` without root privileges (by moving `sudo` out of the CPI and
into a helper script which can be swapped in for a different, sudo-less,
helper script).

[#177401654] Add heavy stemcell support to Ops Manager AWS

Signed-off-by: Brian Upton <bupton@vmware.com>