Revert "`sudo` the one executable, not the entire script"
We've decided to go in another direction - the script itself will
continue to be run with sudo. The change we are reverting does not work.
There is code that configures the vcap user to be able to run the
stemcell-copy script. If the sudo happens inside of the stemcell-copy
script instead, it would require us to allow password-less sudo of `dd`,
which is a less secure configuration.

This reverts commit 328f0b4ad6e1c8fc5f709157d971df752b583ed7.
This reverts commit 7e83b938cc0198525daf37f3438567fc8793d261.