Update instances types chosen for vm_resources
This updates the list of instance types that are chosen from when using
the vm_resources property. The previous list was about one AWS EC2
generation out of date. The new list was chosen from general purpose
instance types that are available across all regions (including US
GovCloud).

This changes the algorithm for choosing the VM type to be based on
preference order. The current list is ordered by pricing as of current
date.

We kept the t2 instance types as they are one of the only 1 CPU options
available.

[#177650543] **Consider** Updating the BOSH AWS CPI `SUPPORTED_VM_TYPES`

Signed-off-by: Brian Upton <bupton@vmware.com>