Allow to configure throughput for gp3 volumes
according to API docs
https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/EC2/Client.html#create_volume-instance_method
it should be possible to specify desired throughput for a gp3 volume to be
created. Also see AWS docs [1]

Optionally allow throughput to be specified as part of
the cloud properties for a disk type.

TODO add tests
TODO this should only be possible if the disk type is 'gp3'

[1] https://awscli.amazonaws.com/v2/documentation/api/latest/reference/ec2/create-volume.html