replace bundle without env var with --without cli param
after updating ruby (and most probably bundler with it)
the BUNDLE_WITHOUT env var stopped working. Somehow the
gems where missing when calling the cpi while running the
create-env. while debugging it was found that using the
`--without` cli param on the other hand had the expected
results. this fixes:

```
/root/.bosh/installations/a00332ec-3a1e-4af9-64c0-56ba1d9c03ea/packages/ruby-3.1/lib/ruby/site_ruby/3.1.0/bundler/definition.rb:484:in `materialize': Could not find pry-byebug-3.9.0, rspec-3.10.0, rake-12.3.3, webmock-3.12.2, byebug-11.1.3, pry-0.13.1, rspec-core-3.10.1, rspec-expectations-3.10.1, rspec-mocks-3.10.2, addressable-2.8.0, crack
-0.4.5, hashdiff-1.0.1, coderay-1.1.3, method_source-1.0.0, rspec-support-3.10.2, diff-lcs-1.4.4, public_suffix-4.0.7 in any of the sources (Bundler::GemNotFound)
```

when running create-env with the new release version