stop testing paravirtual amis and switch to jammy stemcell
we do not publish any pv amis anymore. they got fully
replaced by hvm amis and the tests should be faster
without testing super old cached pv_amis

use new stemcell for integration tests
currently the env var for the stemcell is not set.
thus the tests default to hardcoded ami that are super
old. add download of a current light jammy stemcell to
the task and pipeline to be able to set a current ami
from the extracted stemcell manifest. we shouldn't test
on soon to be deprecated amis.