update aws cli blob
the current blob will break with an error in the packaging
script because the underlying python package used in it is
depending on the deprecated dist-utils.

the contents of the zip changed as well and the containing
folder is now `aws` as compared to the old `aws-bundle`

this fixes:

```
The distutils package is deprecated and slated for removal in Python 3.12. Use setuptools or check PEP 632 for potential alternatives\n  import distutils.sysconfig\nTraceback (most recent call last):\n  File "/var/vcap/data/tmp/tmpsazh365y/virtualenv-13.0.3/virtualenv.py", line 20, in <module>\n    import distutils.sysconfig\nModuleNotFoundError: No module named \'distutils.sysconfig\'
```

update packaging script with new filenames for extracted zip