fix packaging script after blob update
while working the packaging script didn't put the "binary"
into ${BOSH_INSTALL_TARGET} as the old version of the blob.
this made the errands in the release fail with `aws` not found.

the binary is just a symlink to a python script in the
nested install folders. Add the -b parameter to make sure that the
symlinks to the cli scripts are put into ${BOSH_INSTALL_TARGET}.

this also involves renaming the install folder as that was also
named aws if you just used -i ${BOSH_INSTALL_TARGET}.