Configure Windows AMI through pipeline
Previously, there were two Windows-based AMIs hardcoded as defaults into
the integration tests. However, these AMIs have been deleted, so we
needed to configure a more recent AMI instead.

In the interest of making this AMI's usage more clear, we've deleted the
default from the integration test code and now require it to be
configured in the pipeline.

Fixes:
```
  4) Bosh::AwsCloud::CloudV1 manual networking when raw_instance_storage is true root device configuration when root_disk properties are specified and type is specified and AMI is Windows requests root disk with the specified size and type gp3
     Failure/Error: raise Bosh::Clouds::CloudError, "Stemcell does not contain an AMI in region #{@config.aws.region}" unless available_image

     Bosh::Clouds::CloudError:
       Stemcell does not contain an AMI in region us-west-1
```

[#183054448] Bump Ruby releases in OpsManager

Signed-off-by: Brian Upton <bupton@vmware.com>