Add assume role support
The CPI will use assume role authentication if a role_arn is
specified by the user in the cpi.json file.
Also, the integration tests, bats and e2e tests now use an assumed
role to work.
The pipeline has changed to reflect this and now include the roles
to be assumed.

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>
Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>

[#184632801] Add AssumeRole support to bosh-aws-cpi