Fix integration tests
Adds the support for the assume role in the case of the permissions
auditor when running the integration tests.

Adds the handling for the permissions auditor when checking the
permissions of the role that is performing the tests. Previously this
was done by a user, but now it will be done by a role that is assumed.

Fixes the handling of the session token.

[#184632801] Add AssumeRole support to bosh-aws-cpi